Move proxy button code
