+++ proxy_backends/client/form/form.js
@@ -1,3 +1,8 @@
+// Meteor package imports
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
+import { sAlert } from 'meteor/juliancwirko:s-alert';
+
 // Apinf import
 import { ProxyBackends } from '/proxy_backends/collection';
 import { Proxies } from '/proxies/collection';
@@ -108,6 +113,56 @@ Template.proxyBackend.helpers({
       const frontend = URI(proxy.apiUmbrella.url);
 
       return frontend.host();
+    }
+  },
+});
+
+Template.apiProxy.events({
+  'click #delete-proxy-button': () => {
+    /* Function procedure in generic form
+    1. Delete API Backend on proxy (eg. API Umbrella)
+      - call necessary functions by proxy type
+    2. Delete Proxy Backend on Apinf
+    */
+
+    // Get template instance
+    const instance = Template.instance();
+
+    // Get proxyBackend from template data
+    const proxyBackend = instance.data.proxyBackend;
+
+    // Check proxyBackend exists
+    if (proxyBackend) {
+      // Check if proxyBackend type is apiUmbrella & it has id
+      if (proxyBackend.apiUmbrella && proxyBackend.apiUmbrella.id) {
+        const umbrellaBackendId = instance.data.proxyBackend.apiUmbrella.id;
+
+        // Delete API Backend on API Umbrella
+        Meteor.call(
+          'deleteApiBackendOnApiUmbrella',
+          umbrellaBackendId,
+          (deleteError) => {
+            if (deleteError) {
+              sAlert.error(`Delete failed on Umbrella:\n ${deleteError}`);
+            } else {
+              // Publish changes for deleted API Backend on API Umbrella
+              Meteor.call(
+                'publishApiBackendOnApiUmbrella',
+                umbrellaBackendId,
+                (publishError) => {
+                  if (publishError) {
+                    sAlert.error(`Publish failed on Umbrella:\n ${publishError}`);
+                  } else if (proxyBackend._id) { // Check proxyBackend has _id
+                    // Delete proxyBackend from Apinf
+                    ProxyBackends.remove(proxyBackend._id);
+                    sAlert.success('Successfully deleted proxy settings');
+                  }
+                }
+              );
+            }
+          }
+        );
+      }
     }
   },
 });
